Expand assumptions in predicate solving.
Given a collection of assumptions, we expand those assumptions to
include:

1. The <= relations implied by +'s
2. Transitive combinations of <= relation

We don't expand by commutation of +s... that might take a while.

There is a reliance on apply == to types in here, which may play poorly
with refining types. Also, we do this separately for each predicate,
which is definitely inefficient, and probably unnecessary.