Begin fixing treatment of map-arg, laws for map.
This patch combines a couple of steps towards better handling of the
 mapping operators over rows.

1. We now interpret map-arg : R[k -> k'] -> k -> R[k'] uniformly by

     map-arg r = \X. map (\Y. Y X) r

2. We now have a normalization step

    map f (map g r)
        -->
    map (f . g) r

   where, unfortunately, composition is implemented manually and
   there are special cases for the function arrow.

3. I have begun removing predicate solving steps that incorrectly
   assume that f x ~ g y iff f ~ g /\ x ~ y.

As a consequence, we can now typecheck all the examples from the
ICFP'23 paper, albeit with some additional annotations.

As the above points should suggest, there is still a lingering and
significant higher-order unification muddle to be sorted out.