Fix dubious unification cases for singletons.
Previously I had implemented a pretty dubious rule for singletons,
something like

    τ ~ Σ α  ⇒  α  ↦ {τ}

There are multiple problems with this: it can generate ill-formed rows
(when τ is not a labeled type), it is over-eager when τ is an unreduced
type application, and so forth.

The new rule is

    (ℓ := τ) ~ Σ ρ  ⇒  {ℓ := τ} ~ ρ

which I guess builds in some knowledge of the row theory (not a bad
thing) and also avoids the spurious cases above.