Support configuration files
Rosi now reads options from three locations, in this order:

 - XDG_CONFIG/.rosi
 - ./.rosi
 - command line options

The configuration file name is based on the executable name, so if you
symlink a different executable name for Rosi, it'll look for different
configuration files.

There's a `--reset` option which throws away options processed so far.
This can be used to ignore options in the global or local configuration
file.