Generate subrow containments from containment assumptions.
This patch sucks for several reasons:

1. We're generating a decent number of containments, and taking
a long time to do so. For the former, I think a containment that
the full `AllF` is in another row generates ~256 containments. This
is more than we've had to work with heretofore, but not a completely
stupid number. For the latter, this increases type checking time by
100%, which does not seem justified to me.

2. We're generating duplicate predicates, which should not be
possible. To work around this, I've added a brute force filter. I
suspect this is not positively contributing to the runtime.

For the moment, this allows me to make progress on the small step
example, but this needs reconsideration in the future.