First shot at implementing `fold`
This is assuming (and correcting...) the typing
rule from the ICFP '23 paper. There are one real
and one potential gap.

The real gap is that `fold` doesn't know how to
eliminate a `syn`: neither fixes the size of the record.
For the moment, evaluation will just fail.

The potential gap is that there's no type family, as
turns out to be useful for `syn` and `ana`.