Add qualified names
This is basically a first step towards having something resembling
proper namespace support. Top-level definitions are now always
automatically qualified by modules, where the module name is
determined from the path of the file. Names in code can be either
qualified  or unqualified. Unqualified names resolve to the nearest
definition

Module names are determined from file names by including all
directories between the import root (either "." or a "-i" specified
on the command line) and the file. This is consistent with the way
imports work, so it should provide predictable results......

Fixes #10