Track boundedness of evidence.
To see the essence of what's going on here, consider a projection

    prj [z - x, z, VComplLeq v] (every True)

for some evidence variable `v`. Without knowing the size of of `z`,
we have no idea of the size of the resulting record. We know which
fields *not* to include, but there may be arbitrarily many fields.

In the past, I basically ignored this issue entirely. As a result, the
projection above would generate an arbitrarily large record, which
could later cause divergence.

I am still using "unbounded" evidence. I think there is a version that
uses only finite evidence, annotated either as inclusion or exclusion
evidence, but I am leaving that for a later effort.

Fixes #11