Reorganize guessing in predicate solving.
This is essentially a checkpoint towards a proper implementation of
backtracking in predicate solving. For the time being, this patch
separates the guessing logic into individual functions, reorders how
they are called, and moves backtracking logic into `CheckM`. This does
actually make a little bit of progress on typechecking some examples,
although I think that is essentially accidental.